Output to correct display with multiple monitors
In modern multi-monitor setups, a GdkScreen often consists of multiple monitors
that are merged to form a large screen area.  This meant that since the code
was using the GdkScreen width and height to set the window size, it would be
larger than any physical monitor, so when full screen is called, the prompt
would appear on a different physical monitor than the current one.
Furthermore, the pixmap used was always that from the monitor in the top left.

This changes to always output to the monitor of the active window and use the
portion of the pixbuf from that monitor too.